+++ src/main/kotlin/no/nav/klage/oppgave/api/Norg2Controller.kt
@@ -1,7 +1,7 @@
 package no.nav.klage.oppgave.api
 
-import no.nav.klage.oppgave.clients.Norg2Client
-import no.nav.klage.oppgave.domain.norg2.Enhet
+import no.nav.klage.oppgave.clients.norg2.Enhet
+import no.nav.klage.oppgave.clients.norg2.Norg2Client
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RestController
