+++ src/main/kotlin/no/nav/klage/oppgave/api/OppgaveController.kt
@@ -3,12 +3,12 @@ package no.nav.klage.oppgave.api
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
+import no.nav.klage.oppgave.api.view.Oppgave
+import no.nav.klage.oppgave.api.view.OppgaverQueryParams
+import no.nav.klage.oppgave.api.view.OppgaverRespons
 import no.nav.klage.oppgave.config.SecurityConfiguration.Companion.ISSUER_AAD
-import no.nav.klage.oppgave.domain.OppgaverQueryParams
 import no.nav.klage.oppgave.domain.Saksbehandlerfradeling
 import no.nav.klage.oppgave.domain.Saksbehandlertildeling
-import no.nav.klage.oppgave.domain.view.Oppgave
-import no.nav.klage.oppgave.domain.view.OppgaverRespons
 import no.nav.klage.oppgave.exceptions.NotMatchingUserException
 import no.nav.klage.oppgave.exceptions.OppgaveIdWrongFormatException
 import no.nav.klage.oppgave.exceptions.OppgaveVersjonWrongFormatException
@@ -132,5 +132,5 @@ class OppgaveController(
 //        return ResponseEntity.ok().location(uri).body(oppgave)
 //    }
 //
-    
+
 }
