+++ src/main/kotlin/no/nav/klage/oppgave/api/OppgaverQueryParamsMapper.kt
@@ -1,6 +1,6 @@
 package no.nav.klage.oppgave.api
 
-import no.nav.klage.oppgave.domain.OppgaverQueryParams
+import no.nav.klage.oppgave.api.view.OppgaverQueryParams
 import no.nav.klage.oppgave.domain.OppgaverSearchCriteria
 import no.nav.klage.oppgave.exceptions.NotOwnEnhetException
 import no.nav.klage.oppgave.repositories.SaksbehandlerRepository
