+++ src/main/kotlin/no/nav/klage/oppgave/clients/axsys/AxsysClient.kt
@@ -1,4 +1,4 @@
-package no.nav.klage.oppgave.clients
+package no.nav.klage.oppgave.clients.axsys
 
 import brave.Tracer
 import no.nav.klage.oppgave.config.CacheWithRedisConfiguration.Companion.SAKSBEHANDLERE_I_ENHET_CACHE
