+++ src/main/kotlin/no/nav/klage/oppgave/clients/azure/MicrosoftGraphClient.kt
@@ -1,4 +1,4 @@
-package no.nav.klage.oppgave.clients
+package no.nav.klage.oppgave.clients.azure
 
 import no.nav.klage.oppgave.domain.MicrosoftGraphIdentResponse
 import no.nav.klage.oppgave.domain.MicrosoftGraphNameResponse
