+++ src/main/kotlin/no/nav/klage/oppgave/clients/azure/MicrosoftGraphResponse.kt
