+++ src/main/kotlin/no/nav/klage/oppgave/clients/gosys/Oppgave.kt
