+++ src/main/kotlin/no/nav/klage/oppgave/clients/norg2/ArbeidsfordelingRequest.kt
@@ -1,15 +1,13 @@
-package no.nav.klage.oppgave.domain.norg2
+package no.nav.klage.oppgave.clients.norg2
 
-data class ArbeidsfordelingResponse(
+data class ArbeidsfordelingRequest(
     val behandlingstema: String,
     val behandlingstype: String,
     val diskresjonskode: String,
-    val enhetId: Int,
-    val enhetNavn: String,
-    val enhetNr: String,
+    val enhetNummer: String,
     val geografiskOmraade: String,
     val oppgavetype: String,
-    val skalTilLokalkontor: Boolean,
+    val skjermet: Boolean,
     val tema: String,
     val temagruppe: String
 )
