+++ src/main/kotlin/no/nav/klage/oppgave/clients/norg2/ArbeidsfordelingResponse.kt
@@ -1,13 +1,15 @@
-package no.nav.klage.oppgave.domain.norg2
+package no.nav.klage.oppgave.clients.norg2
 
-data class ArbeidsfordelingRequest(
+data class ArbeidsfordelingResponse(
     val behandlingstema: String,
     val behandlingstype: String,
     val diskresjonskode: String,
-    val enhetNummer: String,
+    val enhetId: Int,
+    val enhetNavn: String,
+    val enhetNr: String,
     val geografiskOmraade: String,
     val oppgavetype: String,
-    val skjermet: Boolean,
+    val skalTilLokalkontor: Boolean,
     val tema: String,
     val temagruppe: String
 )
