+++ src/main/kotlin/no/nav/klage/oppgave/clients/norg2/Enhet.kt
@@ -1,0 +1,5 @@
+package no.nav.klage.oppgave.clients.norg2
+
+data class Enhet(
+    val navn: String
+)
