--- src/main/kotlin/no/nav/klage/oppgave/domain/norg2/Enhet.kt
@@ -1,5 +1,0 @@
-package no.nav.klage.oppgave.domain.norg2
-
-data class Enhet(
-    val navn: String
-)
