+++ src/main/kotlin/no/nav/klage/oppgave/service/OppgaveFacade.kt
@@ -1,0 +1,6 @@
+package no.nav.klage.oppgave.service
+
+class OppgaveFacade(val oppgaveService: OppgaveService) {
+    
+
+}
