+++ src/main/kotlin/no/nav/klage/oppgave/service/OppgaveStatistikkService.kt
@@ -1,6 +1,6 @@
 package no.nav.klage.oppgave.service
 
-import no.nav.klage.oppgave.clients.OppgaveClient
+import no.nav.klage.oppgave.clients.gosys.OppgaveClient
 import no.nav.klage.oppgave.domain.OppgaverSearchCriteria
 import no.nav.klage.oppgave.util.getLogger
 import org.springframework.stereotype.Service
