+++ src/main/kotlin/no/nav/klage/oppgave/service/unleash/ByEnhetStrategy.kt
@@ -2,7 +2,7 @@ package no.nav.klage.oppgave.service.unleash
 
 import no.finn.unleash.UnleashContext
 import no.finn.unleash.strategy.Strategy
-import no.nav.klage.oppgave.clients.AxsysClient
+import no.nav.klage.oppgave.clients.axsys.AxsysClient
 import no.nav.klage.oppgave.util.getLogger
 import org.springframework.stereotype.Component
