+++ src/test/kotlin/no/nav/klage/oppgave/clients/OppgaveClientTest.kt
@@ -4,6 +4,7 @@ import brave.Tracer
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
+import no.nav.klage.oppgave.clients.gosys.OppgaveClient
 import no.nav.klage.oppgave.domain.OppgaverSearchCriteria
 import no.nav.klage.oppgave.domain.gosys.Oppgave
 import no.nav.klage.oppgave.domain.gosys.OppgaveResponse
